test: relax validation of TEST_BUILDKIT_IMAGE to reduce direct dependencies
commit 2d124e0ce954727d640fb3de81d8b36615e7b84f added some integration-tests,
and as part of those tests added github.com/distribution/distribution/v3 as a
direct dependency.

github.com/distribution/distribution/v3 is unreleased, and is still receiving
backward breaking changes. While this dependency is already used elsewhere,
it looks like the only use in buildx was for a sanity check of the custom
TEST_BUILDKIT_IMAGE; we can probably relax that check to avoid adding the
direct dependency.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>