bake: support setting the context to null in bake
Setting the context to null in bake will cause the dockerfile to be sent
over stdin to the build and will prevent using the default context of
"." similar to how this would be used on the command line.

When the JSON is printed from bake, it will reproduce the null so
subsequent uses can keep the value.

Signed-off-by: Jonathan A. Sternberg <jonathan.sternberg@docker.com>