build: set the frontend attributes for callfunc earlier
The frontend attributes that were set when `CallFunc` was set were not
set early enough and it prevented buildkit from seeing the `requestid`.

The `requestid` would only be seen by the `LLBBridgeClient` and it was
not set yet when the `Controller.Solve` method was invoked.

This also had a side effect that the linter could not see that it was
being invoked and the image verifier could not know it shouldn't run.

Signed-off-by: Jonathan A. Sternberg <jonathan.sternberg@docker.com>